feat : 20250627 voidraya

+++ backend/src/main/java/com/example/backend/DB/Guardians.java
@@ -19,7 +19,7 @@ public class Guardians {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "login_id", unique = true, nullable = false, length = 50)
     private String loginId;
@@ -27,35 +27,38 @@ public class Guardians {
     @Column(name = "login_pw", nullable = false)
     private String loginPw;
 
-    @Column(name = "guardian_name", nullable = false, length = 50)  // 테이블 스키마에 맞게 길이 수정
+    @Column(name = "guardian_name", nullable = false, length = 50)
     private String guardianName;
 
     @Column(length = 20)
     private String phone;
 
-    @Column(length = 100)  // unique 제거 (테이블 스키마에 없음)
+    @Column(length = 100)
     private String email;
 
-    @Column(length = 30)  // 테이블 스키마에 맞게 길이 수정
+    @Column(length = 30)
     private String relationship;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false, length = 20)
     private Role role;
 
-    @Column(name = "is_active")  // 컬럼명 매핑 추가
+    @Column(name = "is_active")
     @Builder.Default
     private Boolean isActive = true;
 
-    @Column(name = "created_at", nullable = false, updatable = false)  // 핵심: created_at으로 매핑
+    @Column(name = "created_at", nullable = false, updatable = false)
     private LocalDateTime registeredAt;
 
-    @Column(name = "updated_at")  // 컬럼명 매핑 추가
+    @Column(name = "updated_at")
     private LocalDateTime updatedAt;
 
     /**
-     * 엔티티 저장 전 실행되는 메서드
+     * 엔티티 저장 전 실행되는 메서드들은 아래에 작성
      */
+    //========================== 엔티티 저장 전 ==========================
+    // isActive == null -> 신규일 경우에만 해당
+    // 즉 신규로 계정을 생성했을 경우 사용여부는 True로 설정
     @PrePersist
     protected void onCreate() {
         registeredAt = LocalDateTime.now();
@@ -65,8 +68,10 @@ public class Guardians {
     }
 
     /**
-     * 엔티티 업데이트 전 실행되는 메서드
+     * 엔티티 업데이트 전 실행되는 메서드들은 아래에 작성
      */
+    //========================== 엔티티 업데이트 전 ==========================
+    // 생성일자 Update
     @PreUpdate
     protected void onUpdate() {
         updatedAt = LocalDateTime.now();
